pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "codegen_agent"
version = "0.1.0"
description = "Autonomous Codegen Agent"
requires-python = ">=3.8"
dependencies = [
    "pyyaml",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
]

[project.scripts]
codegen-agent = "codegen_agent.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["Legacy Reference", "benchmark_output", "test_output", ".venv", ".git"]
